refactor(lsmkv): compactOnce takes ctx directly; bridge lives in compactOrCleanup
Etienne review on weaviate/weaviate#11461: passing
`cyclemanager.ShouldAbortCallback` to compactOnce forced test callers
into a cryptic `compactOnce(nil)` shape. Inverted the responsibility:

- compactOnce(ctx context.Context) — direct ctx, no callback. Tests
  call compactOnce(context.Background()) for "never abort" or
  compactOnce(cancelledCtx) for "abort immediately".
- compactOrCleanup hosts the shouldAbort → ctx bridge (pre-cancel
  at entry + 50ms-poll watcher goroutine).

All in-tree test sites updated. The integration test now constructs
a pre-cancelled context.WithCancel directly instead of wrapping a
synthetic shouldAbort callback.
