+++ adapters/repos/db/lsmkv/compaction_bench_test.go
@@ -66,7 +66,7 @@ func BenchmarkCompaction(b *testing.B) {
 					require.Equal(b, 2, fileTypes[".db"])
 					b.StartTimer()
 
-					once, err := bu.disk.compactOnce(nil)
+					once, err := bu.disk.compactOnce(context.Background())
 					require.NoError(b, err)
 					require.True(b, once)
