+++ adapters/repos/db/lsmkv/compactor_proplength_test.go
@@ -109,7 +109,7 @@ func TestInvertedNaNPropLength(t *testing.T) {
 
 	ok := false
 	for !ok {
-		ok, err = bucket.disk.compactOnce(nil)
+		ok, err = bucket.disk.compactOnce(context.Background())
 		require.NoError(t, err)
 	}
