+++ adapters/repos/db/lsmkv/segment_bloom_filters_test.go
@@ -505,7 +505,7 @@ func dontPrecomputeBloom(ctx context.Context, t *testing.T, opts []BucketOption)
 			WithSecondaryKey(0, []byte("bonjour2"))))
 		require.NoError(t, b.FlushMemtable())
 
-		compacted, err := b.disk.compactOnce(nil)
+		compacted, err := b.disk.compactOnce(context.Background())
 		require.NoError(t, err)
 		require.True(t, compacted)
 	})
